printf

+++ include/RoutingTable.h
@@ -41,7 +41,7 @@ public:
     RoutingTable();
     RoutingTable(Settings* settings, uint32_t ID);
     vector<RoutingTableStruct>* getRoutingTable();
-    RoutingTableStruct makeStruct(uint32_t to, uint32_t via, uint32_t distance);
+    RoutingTableStruct makeStruct(uint32_t to, uint32_t via, uint8_t distance);
     void addRoutingTableStruct(RoutingTableStruct Entry);
     void setMyIdentifier(int ID);
     int getMyIdentifier() const;
